code review updates. move function out of commands.go. removed a goroutine.
